+++ libs/journeys/ui/src/libs/plausibleHelpers/plausibleHelpers.ts
@@ -16,9 +16,6 @@ import {
   VideoStartEventCreateInput
 } from '../../../__generated__/globalTypes'
 import { ActionFields as Action } from '../action/__generated__/ActionFields'
-import { TreeBlock } from '../block'
-import { BlockFields_StepBlock } from '../block/__generated__/BlockFields'
-import { ActionBlock, isActionBlock } from '../isActionBlock'
 import { PlausibleEvent } from '../transformPlausibleBreakdown/transformPlausibleBreakdown'
 
 interface Props {
@@ -73,16 +70,7 @@ export interface BlockAnalytics {
   event: PlausibleEvent
 }
 
-interface StepAnalytics {
-  blockAnalyticsMap: Record<string, BlockAnalytics>
-  totalActiveEvents: number
-}
-
-interface ActionEvent extends PlausibleEvent {
-  target: string
-}
-
-function generateActionTargetKey(action: Action): string {
+export function generateActionTargetKey(action: Action): string {
   switch (action.__typename) {
     case 'NavigateToBlockAction':
       return action.blockId
@@ -122,71 +110,4 @@ export function reverseKeyify(key: string): {
   target?: string
 } {
   return JSON.parse(key)
-}
-
-export function formatEventKey(from: string, to: string): string {
-  return `${from}->${to}`
-}
-
-export function getBlockEventKey(block?: TreeBlock): string {
-  let key = ''
-
-  if (block == null) {
-    return key
-  }
-
-  if (block?.__typename === 'StepBlock') {
-    // Get a key for Steps that have a next block
-    if (block.nextBlockId != null && block.nextBlockId !== '') {
-      key = formatEventKey(block.id, block.nextBlockId)
-    }
-  }
-
-  // Get a key for actions blocks that have an action
-  if (isActionBlock(block) && block.action != null) {
-    const target = generateActionTargetKey(block.action)
-    key = formatEventKey(block.action.parentBlockId, target)
-  }
-
-  return key
-}
-
-export function isActiveActionEvent(
-  event: PlausibleEvent
-): event is ActionEvent {
-  return event.target != null && event.event !== 'navigatePreviousStep'
-}
-
-export function getStepAnalytics(
-  blocks: Array<TreeBlock<BlockFields_StepBlock> | ActionBlock>,
-  eventMap?: Record<string, PlausibleEvent>
-): StepAnalytics {
-  let totalActiveEvents = 0
-  const blockAnalyticsMap: { [key: string]: BlockAnalytics } = {}
-
-  // Iterate over action blocks to populate blockAnalyticsMap and totalActiveEvents
-  blocks.forEach((block) => {
-    if (block == null) return
-
-    const key = getBlockEventKey(block)
-    const event = eventMap?.[key]
-
-    if (event != null) {
-      totalActiveEvents += event.events
-      blockAnalyticsMap[block.id] = { event, percentOfStepEvents: 0 }
-    }
-  })
-
-  // Calculate
-  blocks.forEach((block) => {
-    const analytics = blockAnalyticsMap[block.id]
-
-    if (analytics == null) return
-
-    const percentage = analytics.event.events / totalActiveEvents
-
-    blockAnalyticsMap[block.id].percentOfStepEvents = percentage
-  })
-
-  return { blockAnalyticsMap, totalActiveEvents }
 }
